third_party/blink/web_tests/fast/events/dispatch-event-being-dispatched-expected.txt

Tests that dispatchEvent raises DISPATCH_REQUEST_ERR if the event being dispatched is already being dispatched.

PASS should have got DISPATCH_REQUEST_ERR EventException
PASS redispatchCustom.wasInvoked is true
PASS should have got DISPATCH_REQUEST_ERR EventException
PASS checkCustom.wasInvoked is true
PASS should have got DISPATCH_REQUEST_ERR EventException
PASS successfullyParsed is true

TEST COMPLETE
